+++ integration/src/test/scala/co/topl/bridge/FailedRedemptionModule.scala
@@ -119,7 +119,7 @@ trait FailedRedemptionModule {
         _ <- IO.println("Generating blocks..")
         _ <- IO.println("sentTxId: " + sentTxId)
         _ <- process
-          .ProcessBuilder(DOCKER_CMD, generateToAddress(6, newAddress): _*)
+          .ProcessBuilder(DOCKER_CMD, generateToAddress(7, newAddress): _*)
           .spawn[IO]
           .use(_.exitValue)
         _ <- EmberClientBuilder