+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for sad path: Bridge did not mint the tokens in time.
 - Support for sad path: User did not redeem the tokens in time.
 - Add new parameter `--topl-blocks-to-recover` to specify the number of blocks to wait for the Topl network to recover.
+- Add new parameter `--btc-confirmation-threshold` to specify the number of confirmations needed for the deposit.
 - Add two new fields to the `StartPeginSessionResponse` message: `minHeight` and `maxHeight`.
 This is needed to know the range of blocks to wait for the redemption.